// gen/ts/src/nama/api/v1/media_pb.d.ts
export enum MediaKind {
  /**
   * @generated from enum value: MEDIA_KIND_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MEDIA_KIND_MOVIE = 1;
   */
  MOVIE = 1,

  /**
   * @generated from enum value: MEDIA_KIND_SHOW = 2;
   */
  SHOW = 2,

  /**
   * @generated from enum value: MEDIA_KIND_SEASON = 3;
   */
  SEASON = 3,

  /**
   * @generated from enum value: MEDIA_KIND_EPISODE = 4;
   */
  EPISODE = 4,
}